Add features to update and delete projects on admin panel
