+++ src/app/admin/page.tsx
@@ -8,7 +8,7 @@ const AdminPage = () => {
     <>
       <div className="fixed min-h-dvh inset-0 -z-10 w-full items-center [background:radial-gradient(125%_125%_at_50%_10%,#000_40%,hsl(var(--accent))_100%)]"></div>
 
-      <div className="flex flex-col items-center justify-center mt-10 mb-10">
+      <div className="flex flex-col items-center justify-center">
         <AdminPanel />
       </div>
     </>
